tweetSearch Update
Updated tweetSearch:
- Blacklist option: to filter out bots that may contain a keyword that
you would like to use for queries.
- Slack notifier: integrated slack-notifier gem for easy messaging to a
web hook URL (default: off // To enable: remove the comments from in
front of line 14, 15, and 157 for this commit)

Changed tweet comparison to compare the tweet’s URL. If the most recent
2 tweet URLs are different from each other, it will post the tweet to
the console. If Slack notifications are enabled, it’ll ‘ping’ the Slack
room.