similarity checker
Added some logic (findMatch method) that checks to see if there is a
similarity between a new tweet and a previously stored tweet (in
compareArray).

Sometimes, tweeters remove the RT and post the tweet as their own so
you would still receive the retweet as if it weren’t a retweet. In this
case, if the findMatch method finds a 50% or stronger similarity, it
will consider the tweet to be a retweet and will not include it into
the slack feed.