Significant changes
Cleaned up some of the code and made an enhancement to NITD (near
identical tweet detection). Explanation to how the script works now:

- Receive tweet(s). Number determined by user when the script initially
loads.
- Check for blacklisted user. If the "tweeter" is on the blacklist,
block the tweet.
- Check for near identical tweet. If the contents of the tweet are at
least 50% similar, block the tweet. If the tweet passes the NITD test,
add the tweet to an array for future comparisons.
- Post the tweet to Slack, and repeat every ~60 (or however long you
want to query) seconds.

This allows the script to become better at detecting retweets the
longer it runs, effectively filtering out most (if not all) tweets we
do not need to see.

Additional to that, it used to only be able to pull one tweet at a
time, so if two tweets were sent nearly back-to-back we wouldn't see
them; however, since this script has evolved over the past 2 weeks to
become more intricate, I've added the ability for the script to pull
more than one tweet at a time. This effectively allows us to not miss a
single tweet, as long as it is a unique tweet and was not tweeted by a
blacklisted user.